Add possibility to use crontabs and extend jobs
At the moment we are a bit restricted to the postion as on the grid5000
machines .local/bin is not by default in the path, we could fix this but
require .bashrc to be created for that.

For now the position got to be in the home directory which is the
regular directory.