Fix persist and backup procedure
There was some bashism I replaced to get rid of the whitepace at the end
of readarray. Otherwise there are some changes in the process off
packing and appending the archives (they were compressed but tar may not
operate on them this way so they are unpacked now and repacked later).
Also included is the addition of the repacking of keys to the root user
in case an old backup is used to reinstatiate a machine in a new
deployment.